Update fragment_accompaniment_menu.xml
For the third accompaniment there are some TODO that are already done --> 
TODO: set the RadioButton text to the accompaniment name using the view model
TODO: set the text to the accompaniment description using the view model
TODO: set the text to the accompaniment price using the view model

I just deleted that lines with the TODO done and give the user/student the opportunity to try it themselves. 

Please, if this is wrong and I didn't see it clear, give me some feedback.
Thanks for your amazing job!